__/requirements/dev.txt
-r base.txt
django-debug-toolbar
django-nose
six
fabric
